refactor: move internal InputUtils outside of the public API (yesman.epicfight.api.client.input -> yesman.epicfight.client.input)
